Refactor to extract services
