+++ .idea/.idea.server/.idea/workspace.xml
@@ -8,11 +8,21 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="b85fda57-099d-42e4-84e1-c9f9fc5d91d7" name="Changes" comment="" />
+    <list default="true" id="b85fda57-099d-42e4-84e1-c9f9fc5d91d7" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.idea/.idea.server/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/.idea.server/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/dotNetRdf.Server.Runner/Program.cs" beforeDir="false" afterPath="$PROJECT_DIR$/dotNetRdf.Server.Runner/Program.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/dotNetRdf.Server.Runner/configuration.ttl" beforeDir="false" afterPath="$PROJECT_DIR$/dotNetRdf.Server.Runner/configuration.ttl" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/dotNetRdf.Server/DotNetRdfExtensions.cs" beforeDir="false" afterPath="$PROJECT_DIR$/dotNetRdf.Server/DotNetRdfExtensions.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/dotNetRdf.Server/IServiceEndpoint.cs" beforeDir="false" afterPath="$PROJECT_DIR$/dotNetRdf.Server/IServiceEndpoint.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/dotNetRdf.Server/SparqlQueryEndpoint.cs" beforeDir="false" afterPath="$PROJECT_DIR$/dotNetRdf.Server/SparqlQueryEndpoint.cs" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="DpaMonitoringSettings">
+    <option name="firstShow" value="false" />
   </component>
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
@@ -25,12 +35,16 @@
   "associatedIndex": 8
 }]]></component>
   <component name="ProjectId" id="2uJ2KUAiRFHMW7OE7VuD1Y7vpKb" />
+  <component name="ProjectLevelVcsManager">
+    <ConfirmationsSetting value="1" id="Add" />
+  </component>
   <component name="ProjectViewState">
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
   <component name="PropertiesComponent"><![CDATA[{
   "keyToString": {
+    ".NET Launch Settings Profile.dotNetRdf.Server.Runner: http.executor": "Run",
     "RunOnceActivity.ShowReadmeOnStart": "true",
     "RunOnceActivity.git.unshallow": "true",
     "git-widget-placeholder": "initial-import",
@@ -39,6 +53,7 @@
     "node.js.selected.package.eslint": "(autodetect)",
     "node.js.selected.package.tslint": "(autodetect)",
     "nodejs_package_manager_path": "npm",
+    "settings.editor.selected.configurable": "preferences.sourceCode.C#",
     "vue.rearranger.settings.migration": "true"
   }
 }]]></component>
@@ -82,7 +97,7 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1741952294412</updated>
-      <workItem from="1741952296723" duration="1486000" />
+      <workItem from="1741952296723" duration="9170000" />
     </task>
     <servers />
   </component>
