+++ dotNetRdf.Server.Runner/Program.cs
@@ -5,10 +5,11 @@ using VDS.RDF.Query.Pull.Configuration;
 ConfigurationLoader.AddObjectFactory(new PullQueryProcessorConfigurationFactory());
 
 var builder = WebApplication.CreateBuilder(args);
+builder.Services.AddDotNetRdfServices();
 
 var app = builder.Build();
 
-app.UseDotNetRdf();
+app.MapDotNetRdfEndpoints();
 
 Console.WriteLine(app.Urls.FirstOrDefault());
