+++ dotNetRdf.Server/IServiceEndpoint.cs
@@ -4,6 +4,5 @@ namespace VDS.RDF.Server;
 
 public interface IServiceEndpoint
 {
-    string Path { get; }
     void Register(WebApplication app);
 }
